// app/lib/interface.ts
export enum UserRole {
  USER = 'USER',
  VIP = 'VIP',
  ADMIN = 'ADMIN'
}

export enum PredictionResult {
  PENDING = 'PENDING',
  WON = 'WON',
  LOST = 'LOST'
}

export enum SubscriptionStatus {
  ACTIVE = 'ACTIVE',
  EXPIRED = 'EXPIRED',
  CANCELLED = 'CANCELLED'
}

export enum PaymentStatus {
  PENDING = 'PENDING',
  SUCCESS = 'SUCCESS',
  FAILED = 'FAILED'
}

export enum SubscriptionPlan {
  MONTHLY = 'MONTHLY',
  YEARLY = 'YEARLY',
  DAILY = 'DAILY',
  WEEKLY = 'WEEKLY'
}

// Base interfaces for common properties
export interface BaseModel {
  id: string;
  createdAt: Date;
  updatedAt: Date;
}


// Main interfaces
export interface User extends BaseModel {
  [x: string]: any; // Allow additional properties
  username: string;
  email: string;
  passwordHash: string;
  role: UserRole;
  predictions?: Prediction[];
  subscriptions?: Subscription[];
  payments?: Payment[];
  emailVerified: boolean;
  location: string;
  notifications?: Notification[];
  blogPosts?: BlogPost[];
}

export interface Prediction {
  [x: string]: any;
  league: string;
  sportType: string;
  homeTeam: string;
  awayTeam: string;
  tip: string;
  analysis?: string;
  odds?: number | string;
  result: PredictionResult;
  publishedAt: Date;
  createdBy?: User;
  userId: string;
  gameType: string;
  league_rel?: League;
  validation?: ValidationRules
}

export interface Subscription extends BaseModel {
  user?: User;
  userId: string;
  plan: SubscriptionPlan;
  status: SubscriptionStatus;
  startedAt: Date;
  expiresAt: Date;
  flutterwavePaymentId?: string;
}

export interface Payment extends BaseModel {
  user?: User;
  userId: string;
  amount: number;
  currency: string;
  provider: string;
  status: PaymentStatus;
  reference: string;
}

export interface League extends BaseModel {
  name: string;
  predictions?: Prediction[];
}

export interface Settings extends BaseModel {
  key: string;
  value: string;
  description?: string;
}

export interface Notification extends BaseModel {
  user?: User;
  userId: string;
  message: string;
  read: boolean;
}



export interface BlogPost {
  [x: string]: any
  title: string;
  content: string;
  author?: User;
  slug: string;
  summary: string;
  image?: string;
  category: string
  publishedAt?: Date;
  tags: string | string[];
  status: 'draft' | 'published' | 'archived';
}
// ...
export interface ValidationRules {
  required?: boolean | string;
  min?: number | string;
  max?: number | string;
  minLength?: number | string;
  maxLength?: number | string;
  pattern?: {
    value: RegExp;
    message: string;
  };
  validate?: (value: unknown) => boolean | string | Promise<boolean | string>;
}